webapps/custom-metrics-app/requirements.txt
google-cloud-monitoring
Flask
numpy
psutil